[FAGSYSTEM-194254] flytte navkontor-kallet høyere i komponent-treet
Autofullfør tekster finner ikke navkontor, så vi flytter logikken
høyere opp slik at den også får med seg oppdateringen.
