+++ src/app/personside/MainLayout.tsx
@@ -16,6 +16,8 @@ import BrukerHarUbesvarteMeldinger from './dialogpanel/BrukerHarUbesvarteMelding
 import { guid } from 'nav-frontend-js-utils';
 import useFeatureToggle from '../../components/featureToggle/useFeatureToggle';
 import { FeatureToggles } from '../../components/featureToggle/toggleIDs';
+import { useAppState, useOnMount } from '../../utils/customHooks';
+import { isNotStarted } from '../../rest/utils/restResource';
 
 const Scrollbar = styled.div`
     overflow-y: auto;
@@ -23,11 +25,24 @@ const Scrollbar = styled.div`
     flex-shrink: 1;
 `;
 
+function useBrukersNavKontor() {
+    const resource = useAppState((state) => state.restResources.brukersNavKontor);
+    const dispatch = useDispatch();
+
+    useOnMount(() => {
+        if (isNotStarted(resource)) {
+            dispatch(resource.actions.fetch);
+        }
+    });
+}
+
 function MainLayout() {
     const UI = useSelector((state: AppState) => state.ui);
     const dispatch = useDispatch();
     const tittelId = useRef(guid());
     const usingSFBackend = useFeatureToggle(FeatureToggles.BrukSalesforceDialoger).isOn ?? false;
+
+    useBrukersNavKontor();
 
     const ekspanderDialogpanelHandler = () => {
         if (!UI.dialogPanel.ekspandert) {
